vision_service/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
PyYAML>=6,<7
opencv-python>=4.10,<5
ultralytics>=8.3,<9
